refactor(structure): modify package structure in order to help understandability of the application by having separation of layers, adding an application layer expresses that there is concretions around the application and having ports explains what are the domain features that need to be implemented, renaming lamdas with handlers, if a main is necessary it should never live in the domain layer, rename the adapters to prefix the implementation.
