--- backend/src/main/java/com/codurance/codurawise/Main.java
@@ -1,7 +1,0 @@
-package com.codurance.codurawise;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
