+++ backend/src/main/java/com/codurance/codurawise/application/Main.java
@@ -1,0 +1,7 @@
+package com.codurance.codurawise.application;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
