+++ backend/src/main/java/com/codurance/codurawise/application/handlers/QueryResourcesByTitleAndTag.java
@@ -1,15 +1,15 @@
-package com.codurance.codurawise.lambdas;
+package com.codurance.codurawise.application.handlers;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
+import com.codurance.codurawise.application.handlers.util.MySqlConnectionProvider;
+import com.codurance.codurawise.application.handlers.util.Response;
+import com.codurance.codurawise.application.repositories.MySQLSearchRepository;
 import com.codurance.codurawise.domain.models.Resource;
+import com.codurance.codurawise.domain.ports.repositories.SearchRepository;
 import com.codurance.codurawise.domain.services.SearchService;
-import com.codurance.codurawise.lambdas.util.MySqlConnectionProvider;
-import com.codurance.codurawise.lambdas.util.Response;
-import com.codurance.codurawise.repos.SearchRepository;
-import com.codurance.codurawise.repos.mysql.SearchMySQLRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -18,7 +18,7 @@ import java.util.List;
 import java.util.Map;
 
 public class QueryResourcesByTitleAndTag implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
-  private static final Logger logger = LoggerFactory.getLogger(com.codurance.codurawise.lambdas.GetResourcesByTag.class);
+  private static final Logger logger = LoggerFactory.getLogger(GetResourcesByTag.class);
   private final SearchService searchService;
   private final Response<Resource> response = new Response<>();
 
@@ -26,7 +26,7 @@ public class QueryResourcesByTitleAndTag implements RequestHandler<APIGatewayPro
     SearchRepository repository;
     try {
       Connection connection = MySqlConnectionProvider.createDatabaseConnection();
-      repository = new SearchMySQLRepository(connection);
+      repository = new MySQLSearchRepository(connection);
     } catch (Exception e) {
       throw new RuntimeException(e);
     }
