+++ backend/src/main/java/com/codurance/codurawise/application/repositories/MySQLTagsRepository.java
@@ -1,22 +1,22 @@
-package com.codurance.codurawise.repos.mysql;
+package com.codurance.codurawise.application.repositories;
 
 import com.codurance.codurawise.domain.models.Tag;
-import com.codurance.codurawise.repos.TagRepository;
+
+import com.codurance.codurawise.application.repositories.mysql.util.PreparedStatementExecutor;
+import com.codurance.codurawise.domain.ports.repositories.TagRepository;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.List;
 
-import static com.codurance.codurawise.repos.mysql.util.PreparedStatementExecutor.executeTagQuery;
-
-public class TagsMySQLRepository implements TagRepository {
+public class MySQLTagsRepository implements TagRepository {
 
   private static final String TAG_TABLE = "Tag";
   private static final String TAG_COLUMN = "Tag";
   private final Connection connection;
 
-  public TagsMySQLRepository(Connection connection) {
+  public MySQLTagsRepository(Connection connection) {
     this.connection = connection;
   }
 
@@ -35,7 +35,7 @@ public class TagsMySQLRepository implements TagRepository {
   }
 
   private List<Tag> runQuery(PreparedStatement preparedStatement) throws SQLException {
-    return executeTagQuery(preparedStatement);
+    return PreparedStatementExecutor.executeTagQuery(preparedStatement);
   }
 
 }
