+++ backend/src/main/java/com/codurance/codurawise/domain/ports/repositories/ResourcesRepository.java
@@ -1,4 +1,4 @@
-package com.codurance.codurawise.repos;
+package com.codurance.codurawise.domain.ports.repositories;
 
 import com.codurance.codurawise.domain.models.Resource;
 
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface ResourcesRepository {
 
-    List<Resource> getAllSortedByAverageRatingAndCreationDate();
+  List<Resource> getAllSortedByAverageRatingAndCreationDate();
 
   List<Resource> getByTagSortedByAverageRatingAndCreationDate(String tag);
 }
