--- backend/src/main/java/com/codurance/codurawise/repos/TagRepository.java
@@ -1,11 +1,0 @@
-package com.codurance.codurawise.repos;
-
-import com.codurance.codurawise.domain.models.Tag;
-
-import java.util.List;
-
-public interface TagRepository {
-
-  List<Tag> getAllTags();
-
-}
