+++ backend/src/test/java/com/codurance/codurawise/domain/services/ResourceServiceTest.java
@@ -1,7 +1,7 @@
 package com.codurance.codurawise.domain.services;
 
 import com.codurance.codurawise.domain.models.Resource;
-import com.codurance.codurawise.repos.ResourcesRepository;
+import com.codurance.codurawise.domain.ports.repositories.ResourcesRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
